ksr/requirements.txt
asgiref==3.11.1
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
mysqlclient==2.2.7
pillow==12.1.0
PyJWT==2.11.0
sqlparse==0.5.5
tzdata==2025.3
gunicorn==21.2.0
python-decouple==3.8
